Interpolate and add weather data to a dataframe

When data and the weather measurements do not perfectly line up, perform a linear interpolation between two weather measurements and merge the results into the provided dataset. Only applies to numerical weather columns (see weather for more details).

add_weather(data, weather, cols = "all", interval = "hour", na_gap = 2,
  quiet = FALSE)

Arguments

data
Dataframe. Data with dates or times to which weather data should be added.

weather
Dataframe. Weather data downloaded with weather which should be interpolated and added to data.

cols
Character. Vector containing the weather columns to add or 'all' for all relevant columns. Note that some measure are omitted because they cannot be linearly interpolated (e.g., wind direction).

interval
What interval is the weather data recorded at? "hour" or "day".

na_gap
How many hours or days (depending on the interval) is it acceptible to skip over when interpolating over NAs (see details).

quiet
Logical. Suppress all messages (including messages regarding missing data, etc.)

Details

Dealing with NA values If there are NAs in the weather data, na_gap can be used to specify a tolerance. For example, a tolerance of 2 with an interval of "hour", means that a two hour gap in data can be interpolated over (i.e. if you have data for 9AM and 11AM, but not 10AM, the data between 9AM and 11AM will be interpolated. If, however, you have 9AM and 12PM, but not 10AM or 11AM, no interpolation will happen and data between 9AM and 12PM will be returned as NA.)

Examples

# Weather data only
head(kamloops)

#> # A tibble: 6 x 35
#>   station_name station_id   prov   lat     lon       date                time
#>          <chr>      <dbl> <fctr> <dbl>   <dbl>     <date>              <dttm>
#> 1   KAMLOOPS A      51423     BC  50.7 -120.45 2016-01-01 2016-01-01 00:00:00
#> 2   KAMLOOPS A      51423     BC  50.7 -120.45 2016-01-01 2016-01-01 01:00:00
#> 3   KAMLOOPS A      51423     BC  50.7 -120.45 2016-01-01 2016-01-01 02:00:00
#> 4   KAMLOOPS A      51423     BC  50.7 -120.45 2016-01-01 2016-01-01 03:00:00
#> 5   KAMLOOPS A      51423     BC  50.7 -120.45 2016-01-01 2016-01-01 04:00:00
#> 6   KAMLOOPS A      51423     BC  50.7 -120.45 2016-01-01 2016-01-01 05:00:00
#> # ... with 28 more variables: year <chr>, month <chr>, day <chr>, hour <chr>,
#> #   qual <chr>, weather <chr>, hmdx <dbl>, hmdx_flag <chr>, pressure <dbl>,
#> #   pressure_flag <chr>, rel_hum <dbl>, rel_hum_flag <chr>, temp <dbl>,
#> #   temp_dew <dbl>, temp_dew_flag <chr>, temp_flag <chr>, visib <dbl>,
#> #   visib_flag <chr>, wind_chill <dbl>, wind_chill_flag <chr>, wind_dir <dbl>,
#> #   wind_dir_flag <chr>, wind_spd <dbl>, wind_spd_flag <chr>, elev <dbl>,
#> #   climat_id <chr>, WMO_id <chr>, TC_id <chr>

# Data about finch observations at RFID feeders in Kamloops, BC
head(finches)

#> # A tibble: 6 x 10
#>    animal_id                time logger_id            species   age   sex
#>       <fctr>              <dttm>    <fctr>              <chr> <chr> <chr>
#> 1 041868FF93 2016-03-01 06:57:42      2300 Mountain Chickadee   AHY     U
#> 2 041868FF93 2016-03-01 06:58:41      2300 Mountain Chickadee   AHY     U
#> 3 041868FF93 2016-03-01 07:07:21      2300 Mountain Chickadee   AHY     U
#> 4 06200003BB 2016-03-01 07:32:34      2400        House Finch    SY     M
#> 5 06200003BB 2016-03-01 07:32:35      2400        House Finch    SY     M
#> 6 06200003BB 2016-03-01 07:32:36      2400        House Finch    SY     M
#> # ... with 4 more variables: site_name <chr>, date <date>, lon <dbl>, lat <dbl>

# Match weather to finches
not_run({
  ## Not run
  finch_weather <- add_weather(data = finches, weather = kamloops)
})