Access Station data downloaded from Environment and Climate Change Canada

This function access the built-in stations data frame. You can update this data frame with stations_dl() which will update the locally stored data.

stations()

Format

A data frame:

prov

Province

station_name

Station name

station_id

Environment Canada's station ID number. Required for downloading station data.

climate_id

Climate ID number

WMO_id

Climate ID number

TC_id

Climate ID number

lat

Latitude of station location in degree decimal format

lon

Longitude of station location in degree decimal format

elev

Elevation of station location in metres

tz

Local timezone excluding any Daylight Savings

interval

Interval of the data measurements ('hour', 'day', 'month')

start

Starting year of data record

end

Ending year of data record

normals

Whether current climate normals are available for that station

normals_1981_2010

Whether 1981-2010 climate normals are available for that station

normals_1971_2000

Whether 1981-2010 climate normals are available for that station

Source

https://climate.weather.gc.ca/index_e.html

Details

You can check when this was last updated with stations_meta().

A dataset containing station information downloaded from Environment and Climate Change Canada. Note that a station may have several station IDs, depending on how the data collection has changed over the years. Station information can be updated by running stations_dl().

Examples

stations()

#> The stations data frame hasn't been updated in over 4 weeks. Consider running `stations_dl()` to update it so you have the most recent stations list available

#> # A tibble: 26,325 x 16
#>    prov  station_name station_id climate_id WMO_id TC_id   lat   lon  elev tz   
#>    <chr> <chr>             <dbl> <chr>       <dbl> <chr> <dbl> <dbl> <dbl> <chr>
#>  1 AB    DAYSLAND           1795 301AR54        NA NA     52.9 -112.  689. Etc/…
#>  2 AB    DAYSLAND           1795 301AR54        NA NA     52.9 -112.  689. Etc/…
#>  3 AB    DAYSLAND           1795 301AR54        NA NA     52.9 -112.  689. Etc/…
#>  4 AB    EDMONTON CO…       1796 301BK03        NA NA     53.6 -114.  671. Etc/…
#>  5 AB    EDMONTON CO…       1796 301BK03        NA NA     53.6 -114.  671. Etc/…
#>  6 AB    EDMONTON CO…       1796 301BK03        NA NA     53.6 -114.  671. Etc/…
#>  7 AB    FLEET              1797 301B6L0        NA NA     52.2 -112.  838. Etc/…
#>  8 AB    FLEET              1797 301B6L0        NA NA     52.2 -112.  838. Etc/…
#>  9 AB    FLEET              1797 301B6L0        NA NA     52.2 -112.  838. Etc/…
#> 10 AB    GOLDEN VALL…       1798 301B8LR        NA NA     53.2 -110.  640  Etc/…
#> # … with 26,315 more rows, and 6 more variables: interval <chr>, start <dbl>,
#> #   end <dbl>, normals <lgl>, normals_1981_2010 <lgl>, normals_1971_2000 <lgl>

stations_meta()

#> $ECCC_modified
#> [1] "2021-05-02 23:31:00 UTC"
#> 
#> $weathercan_modified
#> [1] "2021-06-03"
#> 

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

filter(stations(), interval == "hour", normals == TRUE, prov == "MB")

#> The stations data frame hasn't been updated in over 4 weeks. Consider running `stations_dl()` to update it so you have the most recent stations list available

#> # A tibble: 98 x 16
#>    prov  station_name      station_id climate_id WMO_id TC_id   lat    lon  elev
#>    <chr> <chr>                  <dbl> <chr>       <dbl> <chr> <dbl>  <dbl> <dbl>
#>  1 MB    RIVERS PETTAPIECE       3457 501BMMA        NA NA     50.1 -100.   488.
#>  2 MB    ROSSBURN 4 NORTH        3458 501KE01        NA NA     50.8 -101.   610.
#>  3 MB    ELKHORN 2 EAST          3460 5010QFQ        NA NA     49.9 -101.   498.
#>  4 MB    ALEXANDER               3461 5010033        NA NA     49.8 -100.   427 
#>  5 MB    BALDUR                  3463 5010140        NA NA     49.3  -99.3  450.
#>  6 MB    BELMONT                 3466 5010191        NA NA     49.3  -99.4  482.
#>  7 MB    BINSCARTH               3467 5010216        NA NA     50.6 -101.   526.
#>  8 MB    BRANDON A               3471 5010480     71140 YBR    49.9 -100.   409.
#>  9 MB    BRANDON CDA             3472 5010485        NA NA     49.9 -100.   363.
#> 10 MB    CYPRESS RIVER           3481 5010640        NA NA     49.6  -99.1  374.
#> # … with 88 more rows, and 7 more variables: tz <chr>, interval <chr>,
#> #   start <dbl>, end <dbl>, normals <lgl>, normals_1981_2010 <lgl>,
#> #   normals_1971_2000 <lgl>