Get available stations

This function can be used to download a Station Inventory CSV file from Environment and Climate Change Canada. This is only necessary if the station you're interested was only recently added. The 'stations' data set included in this package contains station data downloaded when the package was last compiled. This function may take a few minutes to run.

stations_dl(url = NULL, skip = NULL, verbose = FALSE, quiet = FALSE)

Arguments

url
Character. Url from which to grab the station information (see details)

skip
Numeric. Number of lines to skip at the beginning of the csv. If NULL, automatically derived.

verbose
Logical. Include progress messages

quiet
Logical. Suppress all messages (including messages regarding missing data, etc.)

Value

A tibble containing station names, station ID codes and dates of operation

Details

URL defaults to ftp://client_climate@ftp.tor.ec.gc.ca/Pub/Get_More_Data_Plus_de_donnees/ Station

Examples

 # Update stations data frame
 s <- stations_dl()

#> Error in curl::curl_fetch_memory(url, handle = handle): response reading failed

 # Use new data frame to search for stations
 stations_search("Winnipeg", stn = s)

#> Error in dplyr::filter(stn, interval %in% !!interval, !is.na(start)): object 's' not found