Get timezone from lat/lon

Get timezone from lat/lon

tz_calc(coords = NULL, lat = NULL, lon = NULL, etc = FALSE)

Arguments

coords
Vector or Data frame. Lat, lon coordinates. Can be a vector for a set of two, or a data frame or matrix for multiple. Provide either coords OR lat and lon

lat
Vector. One or more latitudes, must also provide longitude

lon
Vector. One or more longitudes, must also provide latitude

etc
Logical. Return Etc timezone? Non-daylight savings offset only.

Examples

# Daylight savings
tz_calc(lat = 53.881857, lon = -122.786271)

#> [1] "America/Vancouver"

tz_calc(coords = c(53.881857, -122.786271))

#> [1] "America/Vancouver"

# No daylight savings
tz_calc(lat = 53.881857, lon = -122.786271, etc = TRUE)

#> [1] "Etc/GMT+8"

tz_calc(coords = c(53.881857, -122.786271), etc = TRUE)

#> [1] "Etc/GMT+8"