Download weather data from Environment Canada

Downloads data from Environment Canada for one or more stations. For details and units, see the glossary online http://climate.weather.gc.ca/glossary_e.html.

weather(station_ids, start = NULL, end = NULL, interval = "hour",
  trim = TRUE, format = TRUE, string_as = NA, tz_disp = NULL,
  stn = weathercan::stations,
  url = "http://climate.weather.gc.ca/climate_data/bulk_data_e.html",
  encoding = "UTF-8", list_col = FALSE, verbose = FALSE, quiet = FALSE)

Arguments

station_ids
Numeric/Character. A vector containing the ID(s) of the station(s) you wish to download data from. See the stations data frame or the stations_search function to find IDs.

start
Date/Character. The start date of the data in YYYY-MM-DD format (applies to all stations_ids). Defaults to start of range.

end
Date/Character. The end date of the data in YYYY-MM-DD format (applies to all station_ids). Defaults to end of range.

interval
Character. Interval of the data, one of "hour", "day", "month".

trim
Logical. Trim missing values from the start and end of the weather dataframe.

format
Logical. If TRUE, formats data for immediate use. If FALSE, returns data exactly as downloaded from Environment Canda. Useful for dealing with changes by Environment Canada to the format of data downloads.

string_as
Character. What value to replace character strings in a numeric measurement with. See Details.

tz_disp
Character. What timezone to display times in (must be one of OlsonNames()).

stn
Data frame. The stations data frame to use. Will use the one included in the package unless otherwise specified.

url
Character. Url from which to grab the weather data

encoding
Character. Text encoding for download.

list_col
Logical Defaults to FALSE

verbose
Logical. Include progress messages

quiet
Logical. Suppress all messages (including messages regarding missing data, etc.)

Value

A tibble with station ID, name and weather data.

Details

Data can be returned 'raw' (format = FALSE) or can be formatted. Formatting transforms dates/times to date/time class, renames columns, and converts data to numeric where possible. If character strings are contained in traditionally numeric fields (e.g., weather speed may have values such as "< 30"), they can be replaced with a character specified by `string_as`. The default is NA. Formating also replaces data associated with certain flags with NA (M = Missing).

Start and end date can be specified, but if not, it will default to the start and end date of the range (this could result in downloading a lot of data!).

Times are returned as the Etc/GMT offset timezone corresponding to the location. This doesn't include daylight savings. However, for compatibility with other data sets, timezones can be converted by specifying the desired timezone in `tz_disp`.

Examples

not_run({
  kam <- weather(station_ids = 51423,
                 start = "2016-01-01", end = "2016-02-15")
})

stations_search("Kamloops A$", interval = "hour")

#> # A tibble: 2 x 12
#>     prov station_name station_id climate_id WMO_id  TC_id   lat     lon  elev
#>   <fctr>        <chr>     <fctr>     <fctr> <fctr> <fctr> <dbl>   <dbl> <dbl>
#> 1     BC   KAMLOOPS A       1275    1163780  71887    YKA  50.7 -120.44 345.3
#> 2     BC   KAMLOOPS A      51423    1163781  71887    YKA  50.7 -120.45 345.3
#> # ... with 3 more variables: interval <chr>, start <int>, end <int>

stations_search("Prince George Airport", interval = "hour")

#> # A tibble: 1 x 12
#>     prov               station_name station_id climate_id WMO_id  TC_id   lat
#>   <fctr>                      <chr>     <fctr>     <fctr> <fctr> <fctr> <dbl>
#> 1     BC PRINCE GEORGE AIRPORT AUTO      48248    1096453  71302    VXS 53.89
#> # ... with 5 more variables: lon <dbl>, elev <dbl>, interval <chr>,
#> #   start <int>, end <int>

not_run({
  kam.pg <- weather(station_ids = c(48248, 51423),
                    start = "2016-01-01", end = "2016-02-15")

  library(ggplot2)

  ggplot(data = kam.pg, aes(x = time, y = temp, group = station_name, colour = station_name)) +
         geom_line()
})