Easy downloading of weather data from Environment and Climate Change Canada

weathercan is an R package for simplifying the downloading of Historical Climate Data from the Environment and Climate Change Canada (ECCC) website (https://climate.weather.gc.ca)

Details

Bear in mind that these downloads can be fairly large and performing repeated, large downloads may use up Environment Canada's bandwidth unnecessarily. Try to stick to what you need.

There are four main aspects of this package:

Access stations lists

stations (a data frame listing stations)

stations_search() identify stations by name or proximity to a location

stations_dl() re-download/update stations data

Download weather data

weather_dl()

Merge weather data into other data sets through interpolation over time

weather_interp()

Download climate normals data

normals_dl()

We also include several practice data sets:

finches

kamloops

kamloops_day

pg

As well as several vignettes available in the package and Online:

General Usage: vignette("weathercan", package = "weathercan")

Interpolating: vignette("interpolation", package = "weathercan")

Reproducibility: vignette("reproducibility", package = "weathercan")

Flags and Codes: vignette("flags", package = "weathercan")

Weather: Terms and Units: vignette("glossary", package = "weathercan")

Climate Normals: Terms and Units: vignette("glossary_normals", package = "weathercan")

We also have some advanced articles:

Using weathercan with tidyhydat)

References

Environment and Climate Change Canada: https://www.canada.ca/en/environment-climate-change.html

Glossary of terms https://climate.weather.gc.ca/glossary_e.html

ECCC Historical Climate Data: https://climate.weather.gc.ca/

See also

Author

Maintainer: Steffi LaZerte sel@steffilazerte.ca (ORCID)

Other contributors:

Sam Albers sam.albers@gmail.com (ORCID) [contributor]

Nick Brown nicholas512@gmail.com (ORCID) [contributor]

Kevin Cazelles kevin.cazelles@gmail.com (ORCID) [contributor]