Get available stations

This function can be used to download a Station Inventory CSV file from Environment and Climate Change Canada. This is only necessary if the station you're interested was only recently added. The 'stations' data set included in this package contains station data downloaded when the package was last compiled. This function may take a few minutes to run.

stations_dl(url = NULL, skip = NULL, verbose = FALSE, quiet = FALSE)

Arguments

url
Character. Url from which to grab the station information (see details)

skip
Numeric. Number of lines to skip at the beginning of the csv. If NULL, automatically derived.

verbose
Logical. Include progress messages

quiet
Logical. Suppress all messages (including messages regarding missing data, etc.)

Value

A tibble containing station names, station ID codes and dates of operation

Details

URL defaults to ftp://client_climate@ftp.tor.ec.gc.ca/Pub/Get_More_Data_Plus_de_donnees/ Station

Examples

 # Update stations data frame
 s <- stations_dl()

#> According to Environment Canada, Modified Date: 2018-07-02 01:31 UTC

#> Environment Canada Disclaimers:
#> "Station Inventory Disclaimer: Please note that this inventory list is a snapshot of stations on our website as of the modified date, and may be subject to change without notice."
#> "Station ID Disclaimer: Station IDs are an internal index numbering system and may be subject to change without notice."

 # Use new data frame to search for stations
 stations_search("Winnipeg", stn = s)

#> # A tibble: 84 x 12
#>    prov  station_name      station_id climate_id WMO_id TC_id   lat   lon  elev
#>    <fct> <chr>             <fct>      <fct>      <fct>  <fct> <dbl> <dbl> <dbl>
#>  1 MB    WINNIPEG A CS     27174      502S001    71849  XWG    49.9 -97.2   239
#>  2 MB    WINNIPEG A CS     27174      502S001    71849  XWG    49.9 -97.2   239
#>  3 MB    WINNIPEG A CS     27174      502S001    71849  XWG    49.9 -97.2   239
#>  4 MB    WINNIPEG A FIRE … 27525      5023223    <NA>   <NA>   49.9 -97.2   239
#>  5 MB    WINNIPEG A FIRE … 27525      5023223    <NA>   <NA>   49.9 -97.2   239
#>  6 MB    WINNIPEG ANDERSO… 3689       5023170    <NA>   <NA>   49.9 -97.2    NA
#>  7 MB    WINNIPEG ANDERSO… 3689       5023170    <NA>   <NA>   49.9 -97.2    NA
#>  8 MB    WINNIPEG ASSINIB… 3690       5023174    <NA>   <NA>   49.9 -97.2   229
#>  9 MB    WINNIPEG ASSINIB… 3690       5023174    <NA>   <NA>   49.9 -97.2   229
#> 10 MB    WINNIPEG BLUMBER… 3691       5023178    <NA>   <NA>   49.9 -97.4   236
#> # ... with 74 more rows, and 3 more variables: interval <chr>, start <int>,
#> #   end <int>