Get available stations

This function can be used to download a Station Inventory CSV file from Environment and Climate Change Canada. This is only necessary if the station you're interested was only recently added. The 'stations' data set included in this package contains station data downloaded when the package was last compiled. This function may take a few minutes to run.

stations_dl(skip = NULL, verbose = FALSE, quiet = FALSE)

Arguments

skip
Numeric. Number of lines to skip at the beginning of the csv. If NULL, automatically derived.

verbose
Logical. Include progress messages

quiet
Logical. Suppress all messages (including messages regarding missing data, etc.)

Value

A tibble containing station names, station ID codes, dates of operation, as well as whether or not there are data on climate normals.

Details

The stations list is downloaded from the url stored in the option weathercan.urls.stations. To change this location use options(weathercan.urls.stations = "your_new_url").

The list of which stations have climate normals is downloaded from the url stored in the option weathercan.urls.stations.normals. To change this location use options(weathercan.urls.normals = "your_new_url").

Currently there are two sets of climate normals available: 1981-2010 and 1971-2000. Whether a station has climate normals for a given year range is specified in normals_1981_2010 and normals_1971_2000, respectively.

The column normals represents the most current year range of climate normals (i.e. currently 1981-2010)

Examples

stations_dl,Get available stations,Get available stations — stations_dl,Source: R/stations.R,stations_dl.Rd,character(0),stations_dl,character(0),list(title = "Description", contents = "
This function can be used to download a Station Inventory CSV file from\nEnvironment and Climate Change Canada. This is only necessary if the station\nyou're interested was only recently added. The 'stations' data set included\nin this package contains station data downloaded when the package was last\ncompiled. This function may take a few minutes to run.
"),list(description = "This function can be used to download a Station Inventory CSV file from\nEnvironment and Climate Change Canada. This is only necessary if the station\nyou're interested was only recently added. The 'stations' data set included\nin this package contains station data downloaded when the package was last\ncompiled. This function may take a few minutes to run."),stations_dl(skip = NULL, verbose = FALSE, quiet = FALSE),list(list(name = "skip", description = "
Numeric. Number of lines to skip at the beginning of the csv. If\nNULL, automatically derived.
"), list(name = "verbose", description = "
Logical. Include progress messages
"), list(name = "quiet", description = "
Logical. Suppress all messages (including messages regarding\nmissing data, etc.)
")),TRUE,

# \donttest{
if(check_eccc()) {
  # Update stations data frame
  stations_dl()

  # Updated stations data frame is now automatically used
  stations_search("Winnipeg")
}

#> According to Environment Canada, Modified Date: 2021-05-02 23:31 UTC

#> Stations data saved...
#> Use `stations()` to access most recent version and `stations_meta()` to see when this was last updated

#> # A tibble: 85 x 16
#>    prov  station_name station_id climate_id WMO_id TC_id   lat   lon  elev tz   
#>    <chr> <chr>             <dbl> <chr>       <dbl> <chr> <dbl> <dbl> <dbl> <chr>
#>  1 MB    WINNIPEG A …      27174 502S001     71849 XWG    49.9 -97.2  239. Etc/…
#>  2 MB    WINNIPEG A …      27174 502S001     71849 XWG    49.9 -97.2  239. Etc/…
#>  3 MB    WINNIPEG A …      27174 502S001     71849 XWG    49.9 -97.2  239. Etc/…
#>  4 MB    WINNIPEG A …      27525 5023223        NA NA     49.9 -97.2  239. Etc/…
#>  5 MB    WINNIPEG A …      27525 5023223        NA NA     49.9 -97.2  239. Etc/…
#>  6 MB    WINNIPEG AN…       3689 5023170        NA NA     49.9 -97.2   NA  Etc/…
#>  7 MB    WINNIPEG AN…       3689 5023170        NA NA     49.9 -97.2   NA  Etc/…
#>  8 MB    WINNIPEG AS…       3690 5023174        NA NA     49.9 -97.2  229. Etc/…
#>  9 MB    WINNIPEG AS…       3690 5023174        NA NA     49.9 -97.2  229. Etc/…
#> 10 MB    WINNIPEG BL…       3691 5023178        NA NA     49.9 -97.4  236. Etc/…
#> # … with 75 more rows, and 6 more variables: interval <chr>, start <dbl>,
#> #   end <dbl>, normals <lgl>, normals_1981_2010 <lgl>, normals_1971_2000 <lgl>

# }

,list(list(title = "Value", contents = "
A tibble containing station names, station ID codes, dates of\noperation, as well as whether or not there are data on climate normals.
", slug = "value"), list(title = "Details", contents = "
The stations list is downloaded from the url stored in the option\nweathercan.urls.stations. To change this location use\noptions(weathercan.urls.stations = \"your_new_url\").
\n
The list of which stations have climate normals is downloaded from the url\nstored in the option weathercan.urls.stations.normals. To change this\nlocation use options(weathercan.urls.normals = \"your_new_url\").
\n
Currently there are two sets of climate normals available: 1981-2010 and\n1971-2000. Whether a station has climate normals for a given year range is\nspecified in normals_1981_2010 and normals_1971_2000, respectively.
\n
The column normals represents the most current year range of climate\nnormals (i.e. currently 1981-2010)
", 
    slug = "details")),2021,list(name = "weathercan", version = "0.6.0", authors = "Steffi LaZerte"),list(destination = "dev", mode = "release", version_label = "default", version_tooltip = "Released version", in_dev = FALSE),list(root = "../", title = "weathercan"),list(css = "/home/steffi/Projects/weathercan/pkgdown/extra.css"),list(type = "inverse", left = "
\n  \n    \n     \n  \n
\n
\n  Getting Started\n
\n
\n  Reference\n
\n
\n  Articles\n
", right = "
\n  \n    \n     \n  \n
"),list(.present = TRUE),list(version = "1.6.1"),FALSE