Search for stations by name or location

Returns stations that match the name provided OR which are within dist km of the location provided. This is designed to provide the user with information with which to decide which station to then get weather data from.

stations_search(name = NULL, coords = NULL, dist = 10,
  interval = c("hour", "day", "month"), stn = weathercan::stations,
  verbose = FALSE, quiet = FALSE)

Arguments

name
Character. A vector of length 1 or more with text against which to match. Will match station names that contain all components of name, but they can be in different orders and separated by other text.

coords
Numeric. A vector of length 2 with latitude and longitude of a place to match against. Overrides lat and lon if also provided.

dist
Numeric. Match all stations within this many kilometers of the coords.

interval
Character. Return only stations with data at these intervals. Must be any of "hour", "day", "month".

stn
Data frame. The stations data frame to use. Will use the one included in the package unless otherwise specified.

verbose
Logical. Include progress messages

quiet
Logical. Suppress all messages (including messages regarding missing data, etc.)

Value

Returns a subset of the stations data frame which match the search parameters. If the search was by location, an extra column 'distance' shows the distance in kilometres from the location to the station

Examples

stations_search(name = "Kamloops")

#> # A tibble: 40 x 12
#>      prov         station_name station_id climate_id WMO_id  TC_id   lat
#>    <fctr>                <chr>     <fctr>     <fctr> <fctr> <fctr> <dbl>
#>  1     BC             KAMLOOPS       1274    1163779     NA     NA 50.67
#>  2     BC             KAMLOOPS       1274    1163779     NA     NA 50.67
#>  3     BC           KAMLOOPS A       1275    1163780  71887    YKA 50.70
#>  4     BC           KAMLOOPS A       1275    1163780  71887    YKA 50.70
#>  5     BC           KAMLOOPS A       1275    1163780  71887    YKA 50.70
#>  6     BC           KAMLOOPS A      51423    1163781  71887    YKA 50.70
#>  7     BC           KAMLOOPS A      51423    1163781  71887    YKA 50.70
#>  8     BC KAMLOOPS AFTON MINES       1276    1163790     NA     NA 50.67
#>  9     BC KAMLOOPS AFTON MINES       1276    1163790     NA     NA 50.67
#> 10     BC         KAMLOOPS AUT      42203    1163842  71741    ZKA 50.70
#> # ... with 30 more rows, and 5 more variables: lon <dbl>, elev <dbl>,
#> #   interval <chr>, start <int>, end <int>

stations_search(name = "Kamloops", interval = "hour")

#> # A tibble: 3 x 12
#>     prov station_name station_id climate_id WMO_id  TC_id   lat     lon  elev
#>   <fctr>        <chr>     <fctr>     <fctr> <fctr> <fctr> <dbl>   <dbl> <dbl>
#> 1     BC   KAMLOOPS A       1275    1163780  71887    YKA  50.7 -120.44 345.3
#> 2     BC   KAMLOOPS A      51423    1163781  71887    YKA  50.7 -120.45 345.3
#> 3     BC KAMLOOPS AUT      42203    1163842  71741    ZKA  50.7 -120.44 345.0
#> # ... with 3 more variables: interval <chr>, start <int>, end <int>

stations_search(coords = c(53.915495, -122.739379))

#> # A tibble: 16 x 13
#>      prov               station_name station_id climate_id WMO_id  TC_id   lat
#>    <fctr>                      <chr>     <fctr>     <fctr> <fctr> <fctr> <dbl>
#>  1     BC              PRINCE GEORGE        629    1096435     NA     NA 53.92
#>  2     BC              PRINCE GEORGE        629    1096435     NA     NA 53.92
#>  3     BC     PRINCE GEORGE WESTWOOD        637    1096470     NA     NA 53.90
#>  4     BC     PRINCE GEORGE WESTWOOD        637    1096470     NA     NA 53.90
#>  5     BC  PRINCE GEORGE MASSEY AUTO      50169    1096454     NA    VPX 53.90
#>  6     BC     PR GEORGE FOREMAN FLAT        634    1096460     NA     NA 53.93
#>  7     BC     PR GEORGE FOREMAN FLAT        634    1096460     NA     NA 53.93
#>  8     BC          PRINCE GEORGE STP        636    1096468     NA     NA 53.88
#>  9     BC          PRINCE GEORGE STP        636    1096468     NA     NA 53.88
#> 10     BC            PRINCE GEORGE A        631    1096450  71896    YXS 53.89
#> 11     BC            PRINCE GEORGE A        631    1096450  71896    YXS 53.89
#> 12     BC            PRINCE GEORGE A        631    1096450  71896    YXS 53.89
#> 13     BC PRINCE GEORGE AIRPORT AUTO      48248    1096453  71302    VXS 53.89
#> 14     BC PRINCE GEORGE AIRPORT AUTO      48248    1096453  71302    VXS 53.89
#> 15     BC              PRINCE GEORGE      48370    1096439  71896    YXS 53.88
#> 16     BC              PRINCE GEORGE      48370    1096439  71896    YXS 53.88
#> # ... with 6 more variables: lon <dbl>, elev <dbl>, interval <chr>,
#> #   start <int>, end <int>, distance <dbl>

not_run({
  loc <- ggmap::geocode("Prince George, BC")
  stations_search(coords = loc[c("lat", "lon")])
})