Search for stations by name or location

Returns stations that match the name provided OR which are within dist km of the location provided. This is designed to provide the user with information with which to decide which station to then get weather data from.

stations_search(name = NULL, coords = NULL, dist = 10,
  interval = c("hour", "day", "month"), stn = weathercan::stations,
  verbose = FALSE, quiet = FALSE)

Arguments

name
Character. A vector of length 1 or more with text against which to match. Will match station names that contain all components of name, but they can be in different orders and separated by other text.

coords
Numeric. A vector of length 2 with latitude and longitude of a place to match against. Overrides lat and lon if also provided.

dist
Numeric. Match all stations within this many kilometers of the coords.

interval
Character. Return only stations with data at these intervals. Must be any of "hour", "day", "month".

stn
Data frame. The stations data frame to use. Will use the one included in the package unless otherwise specified.

verbose
Logical. Include progress messages

quiet
Logical. Suppress all messages (including messages regarding missing data, etc.)

Value

Returns a subset of the stations data frame which match the search parameters. If the search was by location, an extra column 'distance' shows the distance in kilometres from the location to the station. If no stations are found withing `dist`, the closest 10 stations are returned.

Details

To search by coordinates, users must make sure they have the sp package installed.

Examples

stations_search(name = "Kamloops")

#> # A tibble: 40 x 12
#>    prov   statio… stati… clima… WMO_… TC_id   lat   lon  elev inte… start   end
#>    <fctr> <chr>   <fctr> <fctr> <fct> <fct> <dbl> <dbl> <dbl> <chr> <int> <int>
#>  1 BC     KAMLOO… 1274   11637… <NA>  <NA>   50.7  -120   379 day    1878  1982
#>  2 BC     KAMLOO… 1274   11637… <NA>  <NA>   50.7  -120   379 month  1878  1982
#>  3 BC     KAMLOO… 1275   11637… 71887 YKA    50.7  -120   345 day    1951  2013
#>  4 BC     KAMLOO… 1275   11637… 71887 YKA    50.7  -120   345 hour   1953  2013
#>  5 BC     KAMLOO… 1275   11637… 71887 YKA    50.7  -120   345 month  1951  2013
#>  6 BC     KAMLOO… 51423  11637… 71887 YKA    50.7  -120   345 day    2013  2018
#>  7 BC     KAMLOO… 51423  11637… 71887 YKA    50.7  -120   345 hour   2013  2018
#>  8 BC     KAMLOO… 1276   11637… <NA>  <NA>   50.7  -120   701 day    1977  1993
#>  9 BC     KAMLOO… 1276   11637… <NA>  <NA>   50.7  -120   701 month  1977  1993
#> 10 BC     KAMLOO… 42203  11638… 71741 ZKA    50.7  -120   345 day    2006  2018
#> # ... with 30 more rows

stations_search(name = "Kamloops", interval = "hour")

#> # A tibble: 3 x 12
#>   prov   statio… statio… clima… WMO_… TC_id   lat   lon  elev inte… start   end
#>   <fctr> <chr>   <fctr>  <fctr> <fct> <fct> <dbl> <dbl> <dbl> <chr> <int> <int>
#> 1 BC     KAMLOO… 1275    11637… 71887 YKA    50.7  -120   345 hour   1953  2013
#> 2 BC     KAMLOO… 51423   11637… 71887 YKA    50.7  -120   345 hour   2013  2018
#> 3 BC     KAMLOO… 42203   11638… 71741 ZKA    50.7  -120   345 hour   2006  2018

stations_search(coords = c(53.915495, -122.739379))

#> # A tibble: 16 x 13
#>    prov   station… stati… clim… WMO_… TC_id   lat   lon  elev inte… start   end
#>    <fctr> <chr>    <fctr> <fct> <fct> <fct> <dbl> <dbl> <dbl> <chr> <int> <int>
#>  1 BC     PRINCE … 629    1096… <NA>  <NA>   53.9  -123   569 day    1956  1957
#>  2 BC     PRINCE … 629    1096… <NA>  <NA>   53.9  -123   569 month  1956  1957
#>  3 BC     PRINCE … 637    1096… <NA>  <NA>   53.9  -123   591 day    1974  1976
#>  4 BC     PRINCE … 637    1096… <NA>  <NA>   53.9  -123   591 month  1974  1976
#>  5 BC     PRINCE … 50169  1096… <NA>  VPX    53.9  -123   601 hour   2012  2018
#>  6 BC     PR GEOR… 634    1096… <NA>  <NA>   53.9  -123   579 day    1962  1966
#>  7 BC     PR GEOR… 634    1096… <NA>  <NA>   53.9  -123   579 month  1962  1966
#>  8 BC     PRINCE … 636    1096… <NA>  <NA>   53.9  -123   579 day    1975  2018
#>  9 BC     PRINCE … 636    1096… <NA>  <NA>   53.9  -123   579 month  1975  2007
#> 10 BC     PRINCE … 631    1096… 71896 YXS    53.9  -123   691 day    1942  2009
#> 11 BC     PRINCE … 631    1096… 71896 YXS    53.9  -123   691 hour   1953  2009
#> 12 BC     PRINCE … 631    1096… 71896 YXS    53.9  -123   691 month  1942  2009
#> 13 BC     PRINCE … 48248  1096… 71302 VXS    53.9  -123   680 day    2009  2018
#> 14 BC     PRINCE … 48248  1096… 71302 VXS    53.9  -123   680 hour   2009  2018
#> 15 BC     PRINCE … 48370  1096… 71896 YXS    53.9  -123   691 day    2009  2018
#> 16 BC     PRINCE … 48370  1096… 71896 YXS    53.9  -123   691 hour   2009  2018
#> # ... with 1 more variable: distance <dbl>

# NOT RUN {
loc <- ggmap::geocode("Prince George, BC")
stations_search(coords = loc[c("lat", "lon")])
# }