Search for stations by name or location

Returns stations that match the name provided OR which are within dist km of the location provided. This is designed to provide the user with information with which to decide which station to then get weather data from.

stations_search(name = NULL, coords = NULL, dist = 10,
  interval = c("hour", "day", "month"), stn = weathercan::stations,
  verbose = FALSE, quiet = FALSE)

Arguments

name
Character. A vector of length 1 or more with text against which to match. Will match station names that contain all components of name, but they can be in different orders and separated by other text.

coords
Numeric. A vector of length 2 with latitude and longitude of a place to match against. Overrides lat and lon if also provided.

dist
Numeric. Match all stations within this many kilometres of the coords.

interval
Character. Return only stations with data at these intervals. Must be any of "hour", "day", "month".

stn
Data frame. The stations data frame to use. Will use the one included in the package unless otherwise specified.

verbose
Logical. Include progress messages

quiet
Logical. Suppress all messages (including messages regarding missing data, etc.)

Value

Returns a subset of the stations data frame which match the search parameters. If the search was by location, an extra column 'distance' shows the distance in kilometres from the location to the station. If no stations are found withing `dist`, the closest 10 stations are returned.

Details

To search by coordinates, users must make sure they have the sp package installed.

Examples

stations_search(name = "Kamloops")

#> # A tibble: 40 x 12
#>    prov  station_name station_id climate_id WMO_id TC_id   lat   lon  elev
#>    <fct> <chr>        <fct>      <fct>      <fct>  <fct> <dbl> <dbl> <dbl>
#>  1 BC    KAMLOOPS     1274       1163779    <NA>   <NA>   50.7 -120.  379.
#>  2 BC    KAMLOOPS     1274       1163779    <NA>   <NA>   50.7 -120.  379.
#>  3 BC    KAMLOOPS A   1275       1163780    71887  YKA    50.7 -120.  345.
#>  4 BC    KAMLOOPS A   1275       1163780    71887  YKA    50.7 -120.  345.
#>  5 BC    KAMLOOPS A   1275       1163780    71887  YKA    50.7 -120.  345.
#>  6 BC    KAMLOOPS A   51423      1163781    71887  YKA    50.7 -120.  345.
#>  7 BC    KAMLOOPS A   51423      1163781    71887  YKA    50.7 -120.  345.
#>  8 BC    KAMLOOPS AF… 1276       1163790    <NA>   <NA>   50.7 -120.  701 
#>  9 BC    KAMLOOPS AF… 1276       1163790    <NA>   <NA>   50.7 -120.  701 
#> 10 BC    KAMLOOPS AUT 42203      1163842    71741  ZKA    50.7 -120.  345 
#> # ... with 30 more rows, and 3 more variables: interval <chr>, start <int>,
#> #   end <int>

stations_search(name = "Kamloops", interval = "hour")

#> # A tibble: 3 x 12
#>   prov  station_name station_id climate_id WMO_id TC_id   lat   lon  elev
#>   <fct> <chr>        <fct>      <fct>      <fct>  <fct> <dbl> <dbl> <dbl>
#> 1 BC    KAMLOOPS A   1275       1163780    71887  YKA    50.7 -120.  345.
#> 2 BC    KAMLOOPS A   51423      1163781    71887  YKA    50.7 -120.  345.
#> 3 BC    KAMLOOPS AUT 42203      1163842    71741  ZKA    50.7 -120.  345 
#> # ... with 3 more variables: interval <chr>, start <int>, end <int>

stations_search(coords = c(53.915495, -122.739379))

#> # A tibble: 17 x 13
#>    prov  station_name station_id climate_id WMO_id TC_id   lat   lon  elev
#>    <fct> <chr>        <fct>      <fct>      <fct>  <fct> <dbl> <dbl> <dbl>
#>  1 BC    PRINCE GEOR… 629        1096435    <NA>   <NA>   53.9 -123.  569.
#>  2 BC    PRINCE GEOR… 629        1096435    <NA>   <NA>   53.9 -123.  569.
#>  3 BC    PRINCE GEOR… 637        1096470    <NA>   <NA>   53.9 -123.  591.
#>  4 BC    PRINCE GEOR… 637        1096470    <NA>   <NA>   53.9 -123.  591.
#>  5 BC    PRINCE GEOR… 50169      1096454    <NA>   VPX    53.9 -123.  601 
#>  6 BC    PRINCE GEOR… 50169      1096454    <NA>   VPX    53.9 -123.  601 
#>  7 BC    PR GEORGE F… 634        1096460    <NA>   <NA>   53.9 -123.  579.
#>  8 BC    PR GEORGE F… 634        1096460    <NA>   <NA>   53.9 -123.  579.
#>  9 BC    PRINCE GEOR… 636        1096468    <NA>   <NA>   53.9 -123.  579 
#> 10 BC    PRINCE GEOR… 636        1096468    <NA>   <NA>   53.9 -123.  579 
#> 11 BC    PRINCE GEOR… 631        1096450    71896  YXS    53.9 -123.  691.
#> 12 BC    PRINCE GEOR… 631        1096450    71896  YXS    53.9 -123.  691.
#> 13 BC    PRINCE GEOR… 631        1096450    71896  YXS    53.9 -123.  691.
#> 14 BC    PRINCE GEOR… 48248      1096453    71302  VXS    53.9 -123.  680 
#> 15 BC    PRINCE GEOR… 48248      1096453    71302  VXS    53.9 -123.  680 
#> 16 BC    PRINCE GEOR… 48370      1096439    71896  YXS    53.9 -123.  691 
#> 17 BC    PRINCE GEOR… 48370      1096439    71896  YXS    53.9 -123.  691 
#> # ... with 4 more variables: interval <chr>, start <int>, end <int>,
#> #   distance <dbl>

loc <- ggmap::geocode("Prince George, BC")

#> Information from URL : http://maps.googleapis.com/maps/api/geocode/json?address=Prince%20George,%20BC&sensor=false

stations_search(coords = loc[c("lat", "lon")])

#> # A tibble: 17 x 13
#>    prov  station_name station_id climate_id WMO_id TC_id   lat   lon  elev
#>    <fct> <chr>        <fct>      <fct>      <fct>  <fct> <dbl> <dbl> <dbl>
#>  1 BC    PRINCE GEOR… 629        1096435    <NA>   <NA>   53.9 -123.  569.
#>  2 BC    PRINCE GEOR… 629        1096435    <NA>   <NA>   53.9 -123.  569.
#>  3 BC    PRINCE GEOR… 637        1096470    <NA>   <NA>   53.9 -123.  591.
#>  4 BC    PRINCE GEOR… 637        1096470    <NA>   <NA>   53.9 -123.  591.
#>  5 BC    PRINCE GEOR… 50169      1096454    <NA>   VPX    53.9 -123.  601 
#>  6 BC    PRINCE GEOR… 50169      1096454    <NA>   VPX    53.9 -123.  601 
#>  7 BC    PRINCE GEOR… 636        1096468    <NA>   <NA>   53.9 -123.  579 
#>  8 BC    PRINCE GEOR… 636        1096468    <NA>   <NA>   53.9 -123.  579 
#>  9 BC    PR GEORGE F… 634        1096460    <NA>   <NA>   53.9 -123.  579.
#> 10 BC    PR GEORGE F… 634        1096460    <NA>   <NA>   53.9 -123.  579.
#> 11 BC    PRINCE GEOR… 631        1096450    71896  YXS    53.9 -123.  691.
#> 12 BC    PRINCE GEOR… 631        1096450    71896  YXS    53.9 -123.  691.
#> 13 BC    PRINCE GEOR… 631        1096450    71896  YXS    53.9 -123.  691.
#> 14 BC    PRINCE GEOR… 48248      1096453    71302  VXS    53.9 -123.  680 
#> 15 BC    PRINCE GEOR… 48248      1096453    71302  VXS    53.9 -123.  680 
#> 16 BC    PRINCE GEOR… 48370      1096439    71896  YXS    53.9 -123.  691 
#> 17 BC    PRINCE GEOR… 48370      1096439    71896  YXS    53.9 -123.  691 
#> # ... with 4 more variables: interval <chr>, start <int>, end <int>,
#> #   distance <dbl>